requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0
RPi.GPIO==0.7.1
Adafruit-DHT==1.4.0
